chore(ci): 优化GitHub Actions工作流配置并增强部署流程
- 为ci和deploy工作流添加contents读取权限提升安全性
- deploy工作流新增workflow_run触发器支持流水线联动
- deploy任务新增workflow_dispatch手动触发输入参数环境变量
- 构建步骤中明确Java和Node.js环境设置及依赖缓存
- 拆分发布任务为publish-backend和publish-frontend，增加缓存配置
- 部署阶段支持拉取GHCR最新镜像，失败时回退源码构建
- 部署过程添加健康检查及启动等待保障服务可用
- docker-compose服务容器增加promtail日志采集标签
- DocumentService增强上传文件名安全检查，防止路径穿越攻击
- 改进文件存储流传输参数保证MinIO SDK兼容性
- 移除JacksonConfig配置，简化依赖和配置管理
- 全局替换com.fasterxml.jackson为tools.jackson库接口调用
- LightRagWebuiProxyFilter及ObservabilityProxyFilter增加代理目标Host校验，拒绝非法请求
- MonitorController中HTTP cookie未设置secure属性改为强制true
- 优化Prometheus与Grafana代理路径处理及内容类型判断逻辑
- pom.xml调整依赖版本及排除多余Java时间模块依赖
- frontend依赖升级katex、markdown-it及部分Vue相关库版本
- README文档格式调整，修正部分字符对齐错误

+++ src/main/java/edu/zjut/traceqa/service/FileStorageService.java
@@ -51,7 +51,7 @@ public class FileStorageService {
             minioClient.putObject(PutObjectArgs.builder()
                     .bucket(cfg.getBucket())
                     .object(object)
-                    .stream(new ByteArrayInputStream(data), data.length, -1)
+                    .stream(new ByteArrayInputStream(data), (long) data.length, -1L)
                     .contentType(contentType == null || contentType.isBlank() ? "image/jpeg" : contentType)
                     .build());
         } catch (Exception e) {
